LayoutTests/http/tests/security/same-origin-css-4-expected.txt


PASS Testing same-origin and MIME behavior for CSS. 
PASS xorigincss3.html should not be loaded, even though it is same-origin 
